feat: move `Payload` defintion into base
