+++ uplink/src/collector/tcpjson.rs
@@ -1,8 +1,6 @@
 use flume::{Receiver, RecvError, Sender};
 use futures_util::SinkExt;
 use log::{debug, error, info};
-use serde::{Deserialize, Serialize};
-use serde_json::Value;
 use thiserror::Error;
 use tokio::net::{TcpListener, TcpStream};
 use tokio::time::{Duration, Sleep};
@@ -10,12 +8,13 @@ use tokio::{select, time};
 use tokio_stream::StreamExt;
 use tokio_util::codec::{Framed, LinesCodec, LinesCodecError};
 
-use std::{collections::HashMap, io, sync::Arc};
 use std::pin::Pin;
+use std::{collections::HashMap, io, sync::Arc};
 
 use super::util::DelayMap;
 use crate::base::actions::{Action, ActionResponse, Error as ActionsError};
-use crate::base::{Buffer, Config, Package, Point, Stream, StreamStatus};
+use crate::base::{Config, Package, Stream, StreamStatus};
+use crate::Payload;
 
 #[derive(Error, Debug)]
 pub enum Error {
@@ -236,47 +235,5 @@ impl Bridge {
 
             }
         }
-    }
-}
-
-// TODO Don't do any deserialization on payload. Read it a Vec<u8> which is in turn a json
-// TODO which cloud will double deserialize (Batch 1st and messages next)
-#[derive(Debug, Serialize, Deserialize)]
-pub struct Payload {
-    #[serde(skip_serializing)]
-    pub stream: String,
-    pub sequence: u32,
-    pub timestamp: u64,
-    #[serde(flatten)]
-    pub payload: Value,
-}
-
-impl Payload {
-    pub fn from_string<S: Into<String>>(input: S) -> Result<Self, Error> {
-        Ok(serde_json::from_str(&input.into())?)
-    }
-}
-
-impl Point for Payload {
-    fn sequence(&self) -> u32 {
-        self.sequence
-    }
-
-    fn timestamp(&self) -> u64 {
-        self.timestamp
-    }
-}
-
-impl Package for Buffer<Payload> {
-    fn topic(&self) -> Arc<String> {
-        self.topic.clone()
-    }
-
-    fn serialize(&self) -> serde_json::Result<Vec<u8>> {
-        serde_json::to_vec(&self.buffer)
-    }
-
-    fn anomalies(&self) -> Option<(String, usize)> {
-        self.anomalies()
     }
 }
